requirements.txt
fastapi
uvicorn[standard]
transformers
datasets
sentence-transformers
faiss-cpu
langchain
pydantic
python-multipart
requests
numpy
pandas
scikit-learn
tqdm
peft
accelerate
bitsandbytes
torch
chromadb
python-dotenv
pytest
sentencepiece
